F4 batch 2: elemental tomes, drop-rate charms, cheap-shop pity, boss rate 10%
Seven more food charms on the unused quest items, extending the batch-1
framework (GF_FOOD_BIT 502-508, QuickStartFoodMask widened to u32):

- Scorched/Frosted/Crackling Tome (BOOK1/2/3): fire/ice/shock immunity.
  Early returns at the top of the three elemental contact handlers in
  collision.c keep the hit as plain knockback + iframes but skip the
  burn/freeze/paralysis state and the elemental damage.
- Tingle's Trophy / Jabber Nut / Carlov's Medal: drop-rate charms read
  in CreateRandomItemDrop. Rupees take a second helping of the flat
  QUICKSTART bumps, kinstones double their decayed weight, hearts jump
  from the clamped 2 to weight 30.
- Broken Sword: every shop price capped at 50 rupees in
  QuickStartGetShopPrice, EXCEPT the heart piece's escalating 50+25n
  price, which stays exempt on purpose.

Supporting fixes: the Carlov Medal's vanilla metadata class (0x3) minted
a free bottle through GiveItem, overridden to 0 under QUICKSTART; all
seven ids join the pastries' item-get-cutscene skip in itemOnGround.c
(same broken presentation path that ate pickups silently); Ezlo receipts
are custom strings 39-45 via the existing n+33 indexing.

Also per the user: QUICKSTART_REGION_BOSS_WAVE_CHANCE cut 20 -> 10, and
the boss-spawn-chance-up charm is deliberately not implemented.

Verified: probe picks a spawned tome off the ground (bit set + receipt
on screen); a 120-kill drop trial went 0 -> 8 hearts and 47 -> 53
kinstones vs baseline; the immunity early-returns, the 50-rupee cap,
and the receipt indexing all confirmed in the built ELF's disassembly;
static invariant checker green (uncommon tier now 27 rows).

+++ src/playerUtils.c
@@ -2432,7 +2432,7 @@ void sub_08079DCC(void) {
 // fast. Applied here because the walk state re-derives speed every frame
 // before this runs, so a multiplicative bump cannot compound - the same
 // reason the slope modifier above it is safe.
-extern u8 QuickStartFoodMask(void);
+extern u32 QuickStartFoodMask(void);
 #define QUICKSTART_FOOD_WALK_SPEED 2
 #endif
